MB-37036: Cap automatic num_reader_threads to 64
Limit the auto-configured number of reader threads in the ep-engine
ExecutorPool to 64. The value 64 is chosen as what is capable of being
tested in our typical QE environments.

Note the user can still set a greater value (up to the limit of 512 in
configuration.json) if they manually specify a value for
num_reader_threads.

Change-Id: I857144d64b238923bec9ffefcd0e32e86bc0cfad
Reviewed-on: http://review.couchbase.org/118447
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>