MB-36971: Return KEY_EEXISTS if stream opaque incorrect
During a takeover we may process messages for a stream with an
incorrect opaque temporarily. This is normal during takeover due to
the way that we disconnect DCP streams. With a recent change we made
the Producer disconnect whenever it receive KEY_ENOENT for a Prepare,
Commit, or Abort. This error could have come from a small window
during takeover though which would not be invalid and should not
cause the Producer to disconnect.

Correct this by returning KEY_EEXISTS when a stream exists for a given
vBucket with a different opaque on the Consumer.

Change-Id: I511aa2108da93dfabde4cbcb500bc52952aa777b
Reviewed-on: http://review.couchbase.org/118560
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>