MB-36948: Pass maxVisibleSeqno to ActiveStream::markDiskSnapshot
Interface only change that allows backfill (disk/memory) code to set a
maxVisibleSeqno.

In this commit, we pass a value of the endSeqno as the maxVisibleSeqno
and push that into the readyQueue (SnapshotMarker), that is the end of
the journey because DcpProducer will not push the value any further.

Change-Id: I05a3a73666b4166e8a7ea2d2632ab2c8219260a9
Reviewed-on: http://review.couchbase.org/118664
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>