MB-23086: Make use of ArenaMalloc stat functions
A number of changes to utilise ArenaMalloc stats.

1) Direct the existing "memory" stats to include arena stats for:

* The engine's arena
* The processes 'global' arena

Note that the ep-engine code is mostly unaware of the individual stats that
the ArenaMalloc code returns (except for ep_testsuite).

2) Direct the existing "allocator" stats key to call
ArenaMalloc::getDetailedStats

3) Update the defragmenter task logging to use the arena's fragmentation
stats and remove deadcode from alloc_hooks

4) MemoryTracker no longer provides any functionality and can be removed.

Other parts of the alloc_hooks will be moved in subsequent changes.

Change-Id: I4f4f1d6f171890dcd71683a67b6e8c9bf064c56a
Reviewed-on: http://review.couchbase.org/119221
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>