MB-37161: Set HCS for Disk Snapshots received from non-MH Producers
When a replica vBucket on a MH node is receiving a disk snapshot from
a pre-MH node, no high completed seqno will be sent. This is expected
(given pre-MH knows nothing of HCS), but the replica should treat this
as a HCS of zero (i.e. it's not possible for anything to have been
committed if we don't yet have SyncWrite support).

As part of this, add an additonal check in
CheckpointManager::createSnapshot() that disk snapshots always have a
HCS set.

Change-Id: I37eb82091385dcf55abee47039411689e0d87a06
Reviewed-on: http://review.couchbase.org/119024
Reviewed-by: Daniel Owen <owend@couchbase.com>
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>