MB-37188: Tolerate seqno ack before topology received
Tests have demonstrated during an upgrade to mad-hatter, as the
DCP streams are recreated to flip to support sync replication once all
nodes are upgraded, a seqno ack may be received prior to the topology
being set.

As this is an upgrade to the version adding durability, the replica HPS
should not yet have been moved by real prepares, but it will be
advanced when the end of a disk snapshot is received. The seqno ack
will therefore only occur when an entire disk snapshot has been
received. If the stream is recreated (i.e., starting a takeover stream)
the consumer will send an initial seqno ack with the current HPS value
upon accepting the stream.

This patch queues these acks for processing once the topology is set
(in the same manner acks from not yet known nodes are already
handled)

Change-Id: I00244b2f5320caf07d8990a0ce1d6b39b4e2397a
Reviewed-on: http://review.couchbase.org/119079
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>