MB-37205: Allow duplicate prepares on Disk snapshots
If a replica gets disconnected then it may have an oustanding prepare
when it reconnects to the active. The active node may have moved on
and may send a disk snapshot with a different (newer) prepare for the
same key. This is valid and the newer prepare needs to overwrite the
older one.

To do this we currently have an AwaitingFirstSnapshotMarker state in
PassiveStream. We process the first snapshot marker and set a seqno
on the replica for which we allow prepares that already exist to be
overwritten by newer ones. This works fine, but we do not deal with
the case where the active node drops the cursor of the replica. If
this happens then the replica will not get disconnected, it will just
receive a disk snapshot in the middle of its stream. This disk snapshot
may have a prepare that should overwrite one that already exists on the
replica.

Remove the AwaitingFirstSnapshotMarker state and replace it with a
simple check that we are about to process a Disk snapshot when
considering allowing prepares to overwrite ones that already exist on
a replica.

Change-Id: I5b0d78e838151fe0c7db13120fc351bdf08066ea
Reviewed-on: http://review.couchbase.org/119097
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>