MB-37209: Set maxVisibleSeqno to highSeqno after off-line upgrade
When creating the vbucket_state and "max_visible_seqno" is not
in the JSON file, initialise maxVisibleSeqno to highSeqno, as on
warmup from upgrade the highSeqno is visible.

Also requires that we change the flusher so that mad-hatter always
flushes the max_visible_seqno, for example if a vbucket only ever
received a prepare, we must commit a max_visible_seqno of 0 so that
if KV restarts it doesn't assume we're in the upgrade case.

Change-Id: I8dba17e619b924afee74eb2308aee4813ac82b49
Reviewed-on: http://review.couchbase.org/119149
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>