MB-37186: EPVBucket - Mark prepare as dirty at commit
Not marking the commit as dirty creates a sequential consistency
failure as you can lose your own write (for a small period of time)
in the following case.

When we commit a prepare we do not mark it as dirty as we just update
the CommittedState of the StoredValue in the HashTable. If this
prepare has already been persisted then it is not dirty. If we then
commit it it is not marked as dirty so it is eligible for eviction.
If we evict it and do a get before we flush it then we will trigger a
BGFetch as the item does not exist in the HashTable. This BGFetch will
either return KEY_ENOENT (if no other committed item exists on disk) or
a stale committed item if it runs and completes before the flush of the
item in the CheckpointManager.

Correct this by marking the StoredValue as dirty when we change the
CommittedState.

Change-Id: Icc83a7460f276419fceb65fb2205249098a1c801
Reviewed-on: http://review.couchbase.org/119213
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>