MB-37099: Add locking to bufferevent methods
Add extra locking to the bufferevent methods (inside libevent) as
ep-engine is dispatching the cookie over to another thread and adding
stats from that thread (which breaks the threading model in the core
where we only touch the cookie (and connection) from the front end
worker thread.

Once the dust settles we should refactor ep-engine to stop pushing
data to the cookie from other threads so that we use the same
threading model as the core to remove the need for extra locking.

Change-Id: I872be5acc0db2b420620703f27a112b162846c2f
Reviewed-on: http://review.couchbase.org/119309
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>