MB-37187: Rethrow a more useful exception in dcp_replicator runloop
Currently if we catch an exception in the dcp_replicator runloop then
we will rethrow it if it was not expected. When we do so we lose both
the exception type and message and rethrow an empty std::exception.

Rethrow a runtime_error exception instead with the message of the
original exception to at least preserve the exception message.

Change-Id: I596dc1c1332fefe7d64b0b7167bf8945a00cd6aa
Reviewed-on: http://review.couchbase.org/119417
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>