Cleanup: Remove support for disable listen
This patch removes the support for disable listen on all of
the server sockets when accept() returns an error with error
code set to EMFILE.

Why? As part of starting up the server we try to bump the
maximum number of file descriptors, and adjust the number
of allowed clients to be below this number (and a fixed
set pool of "reserved" file descriptors). Whenever we
accept a client (and we exceed the limit of connections)
we immediately disconnect the client. This means that
we shouldn't reach the situation where the operating
system returns EMFILE unless the engine suddenly starts
using more files than the reserved pool. By disabling
listen we could ultimately DOS ourself (and ns_server
won't be able to connect to the server).

Its better to simplify the code and logic and always
ignore the error.

Change-Id: I1990fbf711fe19ddce1ec97e76775718c789d68f
Reviewed-on: http://review.couchbase.org/119517
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>