Use std::thread for embedded memcached server
And exit the program if we try to start the embedded
server more than once (as that currently don't work due to
the fact that we've got some global variables in there which
is expected to have the same lifetime as the program. Some rely
on the "atexit" handlers to release resources, some get
initialized by the runtime at startup etc).

Instead of giving weird behavior we'll just print an error
message and exit with a nonzero exit code.

Change-Id: Ib483775cfc6957833ba6656ae787af012ef4c821
Reviewed-on: http://review.couchbase.org/119933
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: James Harrison <james.harrison@couchbase.com>