MB-37145: Make backfills explicitly exit if in incorrect state
Backfills which find the stream in an unexpected (i.e., not
backfilling) state at creation time will now immediately complete.

Prior to this, the backfill would be allowed to scan, but the received
items would be discarded.

This change is largely to make the logic of markDiskSnapshot
consistent; the only notable impact would be a small efficiency
improvement from not scanning items when no longer necessary.

Change-Id: Ia0a0dbd3d7ff279ad953c558b733a7093519dd68
Reviewed-on: http://review.couchbase.org/120094
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>