MB-37433: test item pager: Don't expect pager has run after TMPFAIL
Currently the test expects that something is ejected by the pager
after it gets a TMPFAIL. This may not necessarily happen as memory
may drop back below the watermark before the flusher runs. The test
actually has a fallback to deal with this so remove the failure case
and allow the fallback to attempt to force the pager to run instead.

Change-Id: Ic17387d45ab65fd1b1f6dbdfd84b3419891c56fd
Reviewed-on: http://review.couchbase.org/120151
Reviewed-by: James Harrison <james.harrison@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>