MB-37433: Don't allow checkpoint removal/expelling in item pager test
In the item pager test we can currently drop below the HWM due to
checkpoint removal or expelling. This may prevent the item pager
from running and the test to timeout. Disable expelling for this test
and force it to wait for there to be only a single checkpoint to
ensure that checkpoint removal cannot cause us to drop below the HWM.

Change-Id: Ie126b42b0cef0a621b24071151ce314543ff214b
Reviewed-on: http://review.couchbase.org/120228
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>