MB-37151: Update comments about Ephemeral HCS in tests
The TODOs suggested altering the tests to resolve aborts and commits
through the ADM with timeouts/seqno acks to make the HCS used by
Ephemeral disk snapshots advance in line with the persistent
equivalent.

Realistically the tests are clearer as they are and the HCS is more of
an incidental expectation rather than a key focus of the tests; leaving
them as they are but noting why the HCS differs is probably sufficient
for now, and would still guard against unexpected HCS behaviour in the
future.

Change-Id: Iff16c143ce32b85e20dd6369b6af9a3240c9d755
Reviewed-on: http://review.couchbase.org/120891
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>