MB-26476: Track flusher_todo between flush batches
Currently we do not track flusher_todo between flush batches, we
just reset it to 0 at the end of each batch. This results in the
following scenario. If we have items in the rejectQueue we do not
update the highPersistedSeqno, but we do reset flusher_todo. This
causes an intermittent test failure as a failed flush could cause
us to reset flusher_todo unblocking a test thread which then checks
the highPersistedSeqno.

Track flusher_todo between flusher batches (i.e. size of the
rejectQueue).

Change-Id: Ie63cd6fbbfb7b99154b5f116044cca9ea5392728
Reviewed-on: http://review.couchbase.org/120422
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>