MB-28031: Fix and enable test_dcp_cursor_dropping_backfill
This test was disabled as it began to fail often with the error
"Expected `0' to be greater than or equal to `51' - Incorrect snap
start seqno".

The intent of this test is to test what happens when we get cursor
dropped during a backfill. As such, the test checks that we receive
at least 2 snapshots (1 from the first backfill before the cursor
dropping, 1 from the backfill after). The majority of the time the
test is creating a stream whilst all items still exist in the
CheckpointManager though. When we then get cursor dropped and begin
to stream items again we just stream a single disk snapshot as there
was no outstanding backfill to complete. Correct the test by waiting
for something to be removed from the CheckpointManager before creating
our stream to ensure that we create a backfill.

Change-Id: Iff65d84f56df32bb9528eead9e20bd94e9091702
Reviewed-on: http://review.couchbase.org/120477
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>