MB-37433: Replace test_item_pager with ST version
This test is currently sporadically failing. The reason this happens
is because it is possible for things other than the item pager to
reduce memory below the high watermark and they can run before the
item pager can. If for some reason the item pager does not run during
our initial loading of documents, we attempt to trigger it manually by
dropping the watermarks. This is flawed as triggering the item pager
requires something, an op of some sorts, to attempt to reclaim memory.
This means that should we ever attempt to execute this code then the
test will certainly fail as it does nothing that would trigger the
item pager.

Instead of attempting to fix this test and make it more complicated,
simply remove it and add a single threaded version.

Change-Id: I4b4c6b6cb6befc43cccf52993b7309e9a20b2d69
Reviewed-on: http://review.couchbase.org/120559
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: James Harrison <james.harrison@couchbase.com>