MB-37537: Extend engine api with: get_scope_id
The core checks the access for a command before starting
to execute it. It knows the bucket and it knows the collection
id (as that's encoded in the key), but it does not know which
scope the collection belongs to (and it needs that to perform
the privilege check).

The engine knows the full collection configuration and is able
to map a collection to the owning scope.

Change-Id: Icc2d2e108ae5b91f649367cd027bc682e3a189a8
Reviewed-on: http://review.couchbase.org/120618
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>