MB-37546: Introduce the "backup persistence cursor" in CM
We introduce the concept of "backup persistence cursor" for getting rid
of the (old) broken VBucket::rejectQueue.

The main reason behind the rejectQueue was the way getItemsForCursor()
works:
  1) move the cursor forward
  2) pick-up the item
  3) repeat (1) until we get all (or up to a limit) items for cursor
  4) return items
Given that the cursor has moved, if the flusher fails we cannot rely on
the CM for retrieving again the same items.
So, the rejectQueue was used at flush to re-queueing all the items of a
failed flush-batch and pulling them from the queue when re-attempting
the flush.

With the new backup pcursor we do exactly the same as before in
getItemsForCursor(). But (only for persistence) we register a copy of
the cursor before the cursor move. Then:
  1) if flush succeeds, we just remove the copy
  2) if flush fails, we reset the pcursor to the copy

That allows to rely entirely on the CM for re-attemping the flush after
failure.

In this patch the backup cursor is already used at
EPBucket::flushVBucket. That fixes the issue tracked in this MB (ie, new
persisted vbstate could be lost at re-attempting flush after storage
failure at replica). The fix is covered by test.

@todo in follow-ups:
 - test backup pcursor placed at different queue_op
 - test backup pcursor and reset at Checkpoint deduplication
 - test all pcursor reset code-paths at flush
 - remove all the reject-queue code

Change-Id: Icc03e8c18dd1e5998b384efeff777ffff9699106
Reviewed-on: http://review.couchbase.org/121709
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: James Harrison <james.harrison@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>