Add comment about temp initial items to PagingVisitor::visit
PagingVisitor::visit is called per item visited by the item pager.
Here we decide to skip some items such as prepares or expired items
as we do not want to evict them. We do allow the PagingVisitor to
evict temp initial items. Add a comment explaining why this may be
desirable and the affects of evicting a temp initial item to
PagingVisitor::visit for the next person to run into this.

Change-Id: Ibc0baeb2354f73fdd1d646a0e3ad79c825a5b06b
Reviewed-on: http://review.couchbase.org/120993
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Dave Rigby <daver@couchbase.com>