MB-30044: Add a custom build target to generate generated source files
Clang-tidy needs be aware of all source files that a file its checking
depends on. Otherwise when it uses the clang compiler the checks will
fail. In the kv_engine repo, we generate some files at build time which
currently means that we have to do build to have all the source files we
use on disk. So this patch creates a build target that allows us to just
generate the source files that we would normally generate at build time.

Change-Id: I29a5c83b6a6c2513ac4628f2030efc35cdb53374
Reviewed-on: http://review.couchbase.org/121013
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>