MB-37635: Expect compressor visited to be at least previous value
Based on the comment, we expect the function
wait_for_item_compressor_to_settle to wait for at least 1 more run
of the item compressor. The check is flawed though as it waits until
the num_visited count increments by 1, but it may visit more than 1
item in the time it takes to reach this line. In this case we would
have visited at least 1 item, but the check would timeout.

Change-Id: I1a92e3307fa304d4c758501da497337d40fb1f3d
Reviewed-on: http://review.couchbase.org/121071
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Richard de Mellow <richard.demellow@couchbase.com>
Reviewed-by: James Harrison <james.harrison@couchbase.com>