MB-37637: Remove mem_used check from mem_stats test
This test is currently failing sporadically as it expects mem_used
to have increased or remained the same after BG fetching an item,
when compared to the mem_used before it evicted that item, in reality
it sometimes decreases. mem_used can move for a large variety of
reasons, for example a background task that creates a visitor which
is allocated on the heap, so it is very difficult to assert what
happens to mem_used. Given we check value_size which SHOULD be correct
after the BG fetch also, just remove the mem_used check.

Change-Id: I5a126652c56f178aa26c04411c5561d66a27c8a2
Reviewed-on: http://review.couchbase.org/121072
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: James Harrison <james.harrison@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>