MB-37529: Introduce Partial range locks
Normally, when a caller attempts to acquire a range lock they must
either cover the entire requested seqno range or fail.

However, it is desirable that the stale item deleter be able to run on
*some* of the seqList while backfills are reading a *different* range.
Currently, the stale item deleter tries to lock a seqno range covering
the entire seqList, which will fail if a backfill is in progress.

This patch allow users of range locks to specify whether the requested
range must be met exactly, or that a reduced portion of that range
would be acceptable.

This functionality is not used in this patch; the stale item deleter
will be made to use it in a later patch. Once that is done, stale item
deletion will be able to operate on items before the start of any
backfill range lock.

Change-Id: If630099e9f0889bed2e2684b476a68badc3efac7
Reviewed-on: http://review.couchbase.org/121953
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>