Move test only handling of Commits sent as Mutations to test code
Currently in response.cc (production code) we handle the case where
we are passed an item with queue_op commit to a
MutationConsumerMessage. This is never done in production as we do
not sent queue_op over the wire and the DcpConsumer (which creates
the Item) always assumes that a Mutation has queue_op mutation. Move
this code to dcp_reflection_test.cc (where it is actually required)
as we pass the MutationResponse object directly to the Consumer.

Also, correct the setting of the freqCounterValue to the NRUValue of
the old item. It should be set to the freqCounterValue of the old
item instead.

Change-Id: I0c8563d313b63d72a93eca75e74c4bac0e3cd098
Reviewed-on: http://review.couchbase.org/121076
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>