Create engine_testapp interface library
A complete clean build of KV with CCACHE off failed because
executables using engine_testapp (OBJECT library) were
being built before the library dependencies of engine_testapp,
e.g. default_engine.

This occurs because OBJECT libraries do not propagate their
dependencies.

https://gitlab.kitware.com/cmake/cmake/issues/18090

A fix for this issue is to create a separate INTERFACE
library and use that in executables that want engine_testapp.o

Change-Id: Id5718df426ba5df60d5680165caa3aa67cc4bf63
Reviewed-on: http://review.couchbase.org/121131
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>