MB-37009: Explicitly handle all DCP error codes from a Consumer
DCP consumers can return non-success status codes as a response to a
DCP request from a DCP producer. However, prior to this patch we don't
handle these status codes in an explicit manner and thus sometimes
they're handled incorrectly. Which in some cases caused rebalance bugs
as we would ignore error status codes from the DCP consumer.
See MB-36971 and MB-36794.

This patch explicitly handles these non-success codes for a given op.
To ensure we handle the status code correctly. If it's not explicitly
handled for a given op then we will log an error code or throw for an
unknown opcode. Making it clear in the logs that the DCP consumer and
producer ended up in an unexpected state.

Also add unit tests to defend this in the future.

Change-Id: I39e5a0be0868239494b8db782927d757a4e4e568
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/121189
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>