MB-37546: Early-return if persist-vbstate-only in EPB::flushVBucket
Just return if we have already persisted a new vbstate and it was the
only thing to persist.
Ensure that all the expected post-flush operations are performed.

This is one of the last simplifications in the function. Mainly this
makes clear what part of the function is not executed if the
flush-batch is empty.

Change-Id: I16abd8ae803c43e9f4d27376197ac02c5144bdf2
Reviewed-on: http://review.couchbase.org/121339
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>