MB-37529: Purge over partial range
Allow the stale item deleter to run over a smaller range of seqnos in
the event that there are existing range locks.

Currently, the deleter can only run on a range with an early _end_, it
must start at the requested seqno, to avoid having to iterate the
seqList under the write lock to get an iterator to the correct starting
item.

Change-Id: If3c41fb4d37fa55df9ecb113fd370d61c834f6e6
Reviewed-on: http://review.couchbase.org/121954
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>