MB-37702: Flag connections to disconnect on bucket shutdown
On bucket shutdown we currently do not flag connections as
disconnecting. This means that we can accept a StreamRequest during
bucket shutdown. This is problematic as the StreamRequest may come in
after we have closed all other streams for the connection. If it does
so then it may cause a segfault as the BackfillManager no longer exists.

Fix this bug by flagging connections to disconnect before we close their
streams and delete BackfillManagers. This will prevent any incoming
StreamRequest from being accepted.

Change-Id: I0d53581ce3c5bfee5dbbd42135cd3d5d6f950981
Reviewed-on: http://review.couchbase.org/121421
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>